// frontend/src/api/model/AppointmentStatusesDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AppointmentStatusesDto {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  slug?: string;
  name?: string;
  colorCode?: string;
  isTerminal?: boolean;
  description?: string;
}